## 0.7.5 Very Large Update. So this is my "I should have followed my gut" patch that undoes a major design flaw in my code. I should have never decoupled Headers, Footers, Pages, Sheets, and Scripts from Sites. It should have never been done. I even wrote about it in a blog entry about how I had a bad feeling.
This patch re-couples Headers, Footers, Pages, Sheets, and Scripts to Sites. It reworks a lot of the front login and preview login, as well as some of the plugin login. Blogger will be released with the same patch number with all the changes.

To note in this patch, to migrate the databse manually from the old .7.3 version:
1. Make a DB BACKUP!
2. Make a DiCMS back up!
3. Run the following commands in the database:
```
ALTER TABLE `dicms_css_sheets`
	ADD COLUMN `site_id` BIGINT UNSIGNED NOT NULL AFTER `id`;
UPDATE `dicms_css_sheets` SET `site_id`=1;
ALTER TABLE `dicms_footers`
	ADD COLUMN `site_id` BIGINT UNSIGNED NOT NULL AFTER `id`;
UPDATE `dicms_footers` SET `site_id`=1;
ALTER TABLE `dicms_headers`
	ADD COLUMN `site_id` BIGINT UNSIGNED NOT NULL AFTER `id`;
UPDATE `dicms_headers` SET `site_id`=1;
ALTER TABLE `dicms_js_scripts`
	ADD COLUMN `site_id` BIGINT UNSIGNED NOT NULL AFTER `id`;
UPDATE `dicms_js_scripts` SET `site_id`=1;
ALTER TABLE `dicms_pages`
	ADD COLUMN `site_id` BIGINT UNSIGNED NOT NULL AFTER `id`;
UPDATE `dicms_pages` SET `site_id`=1;
ALTER TABLE `dicms_js_scripts`
	ADD COLUMN `active` TINYINT UNSIGNED NOT NULL DEFAULT '0' AFTER `link_type`;
UPDATE `dicms_js_scripts` SET `active`=1;
ALTER TABLE `dicms_css_sheets`
	ADD COLUMN `active` TINYINT UNSIGNED NOT NULL DEFAULT '0' AFTER `link_type`;
UPDATE `dicms_css_sheets` SET `active`=1;
ALTER TABLE `dicms_js_scripts`
	ADD COLUMN `order_by` TINYINT UNSIGNED NOT NULL DEFAULT '0' AFTER `active`;
ALTER TABLE `dicms_css_sheets`
	ADD COLUMN `order_by` TINYINT UNSIGNED NOT NULL DEFAULT '0' AFTER `active`;
```
4. Make a DiCMS backup file, this is your post file
5. Update the system
6. DROP the DiCMS tables
7. Re-run all migrations
8. Restore from DiCMS

You should now have all your content back.

+++ src/Controllers/PageController.php
@@ -22,7 +22,7 @@ class PageController
         ];
     }
 
-    public function index()
+    public function index(Site $site)
     {
         Gate::authorize('viewAny', Page::class);
         $template =
@@ -34,16 +34,23 @@ class PageController
         {
             $template['buttons']['create'] =
                 [
-                    'link' => DiCMS::dicmsRoute('admin.pages.create'),
+                    'link' => DiCMS::dicmsRoute('admin.pages.create', ['site' => $site->id]),
                     'text' => "<i class='fa fa-plus-square'></i>",
                     'classes' => 'bg-text-primary',
                     'title' => __('dicms::pages.new'),
                 ];
+            $template['buttons']['import'] =
+                [
+                    'link' => DiCMS::dicmsRoute('admin.pages.import.show', ['site' => $site->id]),
+                    'text' => '<i class="fa-solid fa-file-import"></i>',
+                    'classes' => 'link-warning',
+                    'title' => __('dicms::pages.import'),
+                ];
         }
-        return view('dicms::pages.index', compact('template'));
+        return view('dicms::pages.index', compact('template', 'site'));
     }
 
-    public function create()
+    public function create(Site $site)
     {
         Gate::authorize('create', Page::class);
         $template =
@@ -53,23 +60,28 @@ class PageController
                     [
                         'back' =>
                             [
-                                'link' => DiCMS::dicmsRoute('admin.pages.index'),
+                                'link' => DiCMS::dicmsRoute('admin.pages.index', ['site' => $site->id]),
                                 'text' => '<i class="fa-solid fa-rotate-left"></i>',
                                 'classes' => 'text-secondary',
                                 'title' => __('dicms::admin.back'),
                             ]
                     ]
             ];
-        return view('dicms::pages.create', compact('template'));
+        return view('dicms::pages.create', compact('template', 'site'));
     }
 
-    public function store(Request $request)
+    public function store(Request $request, Site $site)
     {
         Gate::authorize('create', Page::class);
-        $currentSite = Site::currentSite();
         $data = $request->validate(
             [
-                'name' => 'required|max:255|unique:' . config('dicms.table_prefix') . 'pages',
+                'name' =>
+                    [
+                        'required',
+                        'max:255',
+                        Rule::unique(config('dicms.table_prefix') . 'pages')
+                            ->where(function ($query) use ($site) { $query->where('site_id', $site->id); })
+                    ],
                 'title' => 'nullable',
                 'slug' => 'required|max:255',
                 'path' => 'nullable',
@@ -78,13 +90,19 @@ class PageController
         $data['url'] = (isset($data['path'])? $data['path'] . "/": '' ) . $data['slug'];
         Validator::make(['url' => $data['url']],
         [
-            'url' => 'required|unique:' . config('dicms.table_prefix') . 'pages',
+            'url' =>
+                [
+                    'required',
+                    Rule::unique(config('dicms.table_prefix') . 'pages')
+                        ->where(function ($query) use ($site) { $query->where('site_id', $site->id); })
+                ],
         ], $this->errors())->validate();
 
         $page = new Page();
         $page->fill($data);
+        $page->site_id = $site->id;
         $page->save();
-        return redirect(DiCMS::dicmsRoute('admin.pages.show', ['page' => $page->id]))
+        return redirect(DiCMS::dicmsRoute('admin.pages.show', ['page' => $page->id, 'site' => $site->id]))
             ->with('success-status', __('dicms::pages.success.created'));
     }
 
@@ -96,12 +114,11 @@ class PageController
                 'title' => $page->name,
                 'buttons' => []
             ];
-        $currentSite = Site::currentSite();
-        if(Gate::allows('preview', $currentSite))
+        if(Gate::allows('preview', $page->site))
         {
             $template['buttons']['preview']  =
                 [
-                    'link' => DiCMS::dicmsRoute('admin.preview.home', ['path' => $page->url]),
+                    'link' => DiCMS::dicmsRoute('admin.preview', ['site' => $page->site_id, 'path' => $page->url]),
                     'text' => "<i class='fa-solid fa-eye'></i>",
                     'classes' => 'text-info',
                     'title' => __('dicms::sites.preview_site'),
@@ -111,33 +128,40 @@ class PageController
         {
             $template['buttons']['edit']  =
                 [
-                    'link' => DiCMS::dicmsRoute('admin.pages.edit', ['page' => $page->id]),
+                    'link' => DiCMS::dicmsRoute('admin.pages.edit', ['site' => $page->site_id, 'page' => $page->id]),
                     'text' => "<i class='fa-solid fa-gear'></i>",
                     'classes' => 'text-primary',
                     'title' => __('dicms::pages.edit'),
                 ];
             $template['buttons']['metadata']  =
                 [
-                    'link' => DiCMS::dicmsRoute('admin.pages.metadata', ['page' => $page->id]),
+                    'link' => DiCMS::dicmsRoute('admin.pages.metadata', ['site' => $page->site_id, 'page' => $page->id]),
                     'text' => "<i class='fa-solid fa-info'></i>",
                     'classes' => 'text-primary',
                     'title' => __('dicms::pages.metadata'),
                 ];
+            $template['buttons']['manage'] =
+                [
+                    'link' => DiCMS::dicmsRoute('admin.pages.index', ['site' => $page->site_id]),
+                    'text' => '<i class="fa-solid fa-bars-progress"></i>',
+                    'classes' => 'text-secondary',
+                    'title' => __('dicms::sites.page_management'),
+                ];
         }
-        $template['buttons']['manage'] =
-            [
-                'link' => DiCMS::dicmsRoute('admin.pages.index'),
-                'text' => '<i class="fa-solid fa-bars-progress"></i>',
-                'classes' => 'text-secondary',
-                'title' => __('dicms::sites.page_management'),
-            ];
         $objEditable = $page;
-        return view('dicms::pages.show', compact('template', 'page', 'objEditable'));
+        //is this a plugin page? Then try to get the active site.
+        if($page->plugin_page)
+            $site = Site::activeSite();
+        else //site is the page's site.
+            $site = $page->site;
+        return view('dicms::pages.show', compact('template', 'page', 'objEditable', 'site'));
     }
 
     public function edit(Page $page)
     {
         Gate::authorize('update', $page);
+        if($page->plugin_page)
+            abort(403);
         $template =
             [
                 'title' => __('dicms::pages.edit'),
@@ -145,7 +169,7 @@ class PageController
                     [
                         'back' =>
                             [
-                                'link' => DiCMS::dicmsRoute('admin.pages.show', ['page' => $page->id]),
+                                'link' => DiCMS::dicmsRoute('admin.pages.show', ['page' => $page->id, 'site' => $page->site_id]),
                                 'text' => '<i class="fa-solid fa-rotate-left"></i>',
                                 'classes' => 'text-secondary',
                                 'title' => __('dicms::admin.back'),
@@ -170,16 +194,25 @@ class PageController
             ], $this->errors());
         $page->fill($data);
         $page->save();
-        return redirect(DiCMS::dicmsRoute('admin.pages.show', ['page' => $page->id]))
+        return redirect()->back()
             ->with('success-status', __('dicms::pages.success.updated'));
     }
 
     public function updateSettings(Request $request, Page $page)
     {
         Gate::authorize('update', $page);
+        if($page->plugin_page)
+            abort(403);
         $data = $request->validate(
             [
-                'name' => ['required', 'max:255', Rule::unique(config('dicms.table_prefix') . 'pages')->ignore($page)],
+                'name' =>
+                    [
+                        'required',
+                        'max:255',
+                        Rule::unique(config('dicms.table_prefix') . 'pages')
+                            ->where(function ($query) use ($page) { $query->where('site_id', $page->site_id); })
+                            ->ignore($page)
+                    ],
                 'title' => 'nullable',
                 'slug' => 'required|max:255',
                 'path' => 'nullable',
@@ -187,7 +220,13 @@ class PageController
         $data['url'] = (isset($data['path'])? $data['path'] . "/": '' ) . $data['slug'];
         Validator::make(['url' => $data['url']],
             [
-                'url' => ['required', Rule::unique(config('dicms.table_prefix') . 'pages')->ignore($page)],
+                'url' =>
+                    [
+                        'required',
+                        Rule::unique(config('dicms.table_prefix') . 'pages')
+                            ->where(function ($query) use ($page) { $query->where('site_id', $page->site_id); })
+                            ->ignore($page)
+                    ],
             ], $this->errors())->validate();
         $page->fill($data);
         $page->save();
@@ -198,6 +237,8 @@ class PageController
     public function publishPage(Request $request, Page $page)
     {
         Gate::authorize('publish', $page);
+        if($page->plugin_page)
+            abort(403);
         $page->published = true;
         $page->save();
         return redirect(DiCMS::dicmsRoute('admin.pages.show', ['page' => $page->id]))
@@ -207,6 +248,8 @@ class PageController
     public function unpublishPage(Request $request, Page $page)
     {
         Gate::authorize('activate', $page);
+        if($page->plugin_page)
+            abort(403);
         $page->published = false;
         $page->save();
         return redirect(DiCMS::dicmsRoute('admin.pages.show', ['page' => $page->id]))
@@ -216,6 +259,8 @@ class PageController
     public function duplicatePage(Request $request, Page $page)
     {
         Gate::authorize('create', Page::class);
+        if($page->plugin_page)
+            abort(403);
         $dupe = $page->dupe();
         return redirect(DiCMS::dicmsRoute('admin.pages.edit', ['page' => $dupe->id]))
             ->with('success-status', __('dicms::pages.success.created'));
@@ -224,13 +269,17 @@ class PageController
     public function destroy(Page $page)
     {
         Gate::authorize('delete', $page);
+        if($page->plugin_page)
+            abort(403);
         $page->delete();
-        return redirect(DiCMS::dicmsRoute('admin.pages.index'));
+        return redirect(DiCMS::dicmsRoute('admin.pages.index', ['site' => $page->site_id]));
     }
 
     public function editMetadata(Page $page)
     {
         Gate::authorize('update', $page);
+        if($page->plugin_page)
+            abort(403);
         $template =
             [
                 'title' => __('dicms::pages.metadata'),
@@ -247,5 +296,42 @@ class PageController
             ];
         $obj = $page;
         return view('dicms::sites.metadata', compact('obj', 'template'));
+    }
+
+    public function import(Site $site)
+    {
+        Gate::authorize('create', Page::class);
+        $template =
+            [
+                'title' => __('dicms::headers.import'),
+                'buttons' =>
+                    [
+                        'back' =>
+                            [
+                                'link' => DiCMS::dicmsRoute('admin.headers.index', ['site' => $site->id]),
+                                'text' => '<i class="fa-solid fa-rotate-left"></i>',
+                                'classes' => 'text-secondary',
+                                'title' => __('dicms::admin.back'),
+                            ]
+                    ]
+            ];
+        return view('dicms::pages.import', compact('template', 'site'));
+    }
+
+    public function doImport(Request $request, Site $site)
+    {
+        Gate::authorize('create', Page::class);
+        $data = $request->validate(
+            [
+                'pages' => 'required|array',
+            ], $this->errors());
+        foreach($data['pages'] as $page_id)
+        {
+            $page = Page::find($page_id);
+            if($page)
+                $page->dupe($site);
+        }
+        return redirect(DiCMS::dicmsRoute('admin.pages.index', ['site' => $site->id]))
+            ->with('success-status', __('dicms::pages.success.imported'));
     }
 }
